Move core views mixins to their own file
