+++ com/views.py
@@ -27,10 +27,7 @@ from smtplib import SMTPRecipientsRefused
 from typing import Any
 
 from django.conf import settings
-from django.contrib.auth.mixins import (
-    AccessMixin,
-    PermissionRequiredMixin,
-)
+from django.contrib.auth.mixins import AccessMixin, PermissionRequiredMixin
 from django.core.exceptions import PermissionDenied, ValidationError
 from django.db.models import Max
 from django.forms.models import modelform_factory
@@ -48,12 +45,8 @@ from com.calendar import IcsCalendar
 from com.forms import NewsDateForm, NewsForm, PosterForm
 from com.models import News, NewsDate, Poster, Screen, Sith, Weekmail, WeekmailArticle
 from core.models import User
-from core.views import (
-    CanEditPropMixin,
-    CanViewMixin,
-    QuickNotifMixin,
-    TabedViewMixin,
-)
+from core.views import CanEditPropMixin, CanViewMixin, QuickNotifMixin, TabedViewMixin
+from core.views.mixins import QuickNotifMixin, TabedViewMixin
 from core.views.widgets.markdown import MarkdownInput
 
 # Sith object
