+++ trombi/views.py
@@ -39,15 +39,9 @@ from django.views.generic.edit import CreateView, DeleteView, UpdateView
 
 from club.models import Club
 from core.models import User
-from core.views import (
-    CanCreateMixin,
-    CanEditMixin,
-    CanEditPropMixin,
-    CanViewMixin,
-    QuickNotifMixin,
-    TabedViewMixin,
-)
+from core.views import CanCreateMixin, CanEditMixin, CanEditPropMixin, CanViewMixin
 from core.views.forms import SelectDate
+from core.views.mixins import QuickNotifMixin, TabedViewMixin
 from core.views.widgets.select import AutoCompleteSelectUser
 from trombi.models import Trombi, TrombiClubMembership, TrombiComment, TrombiUser
